On set title also provide searchability
