+++ nucliadb_utils/nucliadb_utils/exceptions.py
@@ -19,5 +19,13 @@
 #
 
 
+class SendToProcessError(Exception):
+    pass
+
+
+class LimitsExceededError(Exception):
+    pass
+
+
 class ShardsNotFound(Exception):
     pass
