+++ nucliadb_writer/nucliadb_writer/api/v1/resource.py
@@ -32,6 +32,7 @@ from nucliadb_models.processing import PushPayload, Source
 from nucliadb_models.resource import NucliaDBRoles
 from nucliadb_telemetry.utils import set_info_on_span
 from nucliadb_utils.authentication import requires
+from nucliadb_utils.exceptions import LimitsExceededError
 from nucliadb_utils.utilities import (
     get_cache,
     get_ingest,
@@ -51,7 +52,6 @@ from nucliadb_writer.api.v1.router import (
     RESOURCES_PREFIX,
     api,
 )
-from nucliadb_writer.exceptions import LimitsExceededError
 from nucliadb_writer.resource.audit import parse_audit
 from nucliadb_writer.resource.basic import (
     parse_basic,
