+++ nucliadb_writer/nucliadb_writer/exceptions.py
@@ -17,12 +17,6 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
-class SendToProcessError(Exception):
-    pass
-
-
-class LimitsExceededError(Exception):
-    pass
 
 
 class ConflictError(Exception):
